+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/_internal/configureDependencies.kt
@@ -3,11 +3,11 @@ package com.javiersc.hubdle.project.extensions._internal
 import com.javiersc.hubdle.project.extensions.apis.HubdleConfigurableExtension
 import com.javiersc.hubdle.project.extensions.config.testing.HubdleConfigTestingExtension.Options
 import com.javiersc.hubdle.project.extensions.config.testing.hubdleTesting
-import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_kotlinTest
-import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_kotlinTestAnnotationsCommon
-import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_kotlinTestJUnit
-import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_kotlinTestJUnit5
-import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_kotlinTestNG
+import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_test
+import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_test_annotations_common
+import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_test_junit
+import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_test_junit5
+import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlin_test_testng
 import com.javiersc.hubdle.project.extensions.kotlin._internal.forKotlinSetsDependencies
 import org.gradle.api.artifacts.MinimalExternalModuleDependency
 import org.gradle.api.provider.Provider
@@ -20,8 +20,8 @@ internal fun HubdleConfigurableExtension.configurableDependencies() {
             TEST_FUNCTIONAL,
             TEST_INTEGRATION,
         ) {
-            implementation(library(jetbrains_kotlin_kotlinTest))
-            implementation(library(jetbrains_kotlin_kotlinTestAnnotationsCommon))
+            implementation(library(jetbrains_kotlin_test))
+            implementation(library(jetbrains_kotlin_test_annotations_common))
         }
 
         project.forKotlinSetsDependencies(
@@ -33,10 +33,10 @@ internal fun HubdleConfigurableExtension.configurableDependencies() {
             if (hubdleTesting.isFullEnabled.get()) {
                 val testModuleFramework: Provider<MinimalExternalModuleDependency?> = provider {
                     when (hubdleTesting.options.get()) {
-                        Options.JUnit -> library(jetbrains_kotlin_kotlinTestJUnit)
-                        Options.JUnitPlatform -> library(jetbrains_kotlin_kotlinTestJUnit5)
-                        Options.TestNG -> library(jetbrains_kotlin_kotlinTestNG)
-                        else -> library(jetbrains_kotlin_kotlinTestJUnit)
+                        Options.JUnit -> library(jetbrains_kotlin_test_junit)
+                        Options.JUnitPlatform -> library(jetbrains_kotlin_test_junit5)
+                        Options.TestNG -> library(jetbrains_kotlin_test_testng)
+                        else -> library(jetbrains_kotlin_test_junit)
                     }.get()
                 }
                 implementation(testModuleFramework)