+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/features/shared/HubdleKotlinCoroutinesFeatureExtension.kt
@@ -16,9 +16,9 @@ import com.javiersc.hubdle.project.extensions.apis.BaseHubdleExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleConfigurableExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.project.extensions.apis.enableAndExecute
-import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlinx_kotlinxCoroutinesAndroid
-import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlinx_kotlinxCoroutinesCore
-import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlinx_kotlinxCoroutinesTest
+import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlinx_coroutines_android
+import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlinx_coroutines_core
+import com.javiersc.hubdle.project.extensions.dependencies._internal.aliases.jetbrains_kotlinx_coroutines_test
 import com.javiersc.hubdle.project.extensions.kotlin._internal.forKotlinSetsDependencies
 import com.javiersc.hubdle.project.extensions.kotlin.android.hubdleAndroidAny
 import com.javiersc.hubdle.project.extensions.kotlin.hubdleKotlinAny
@@ -47,7 +47,7 @@ constructor(
         configurable {
             configure<KotlinProjectExtension> {
                 forKotlinSetsDependencies(MAIN, COMMON_MAIN) {
-                    implementation(library(jetbrains_kotlinx_kotlinxCoroutinesCore))
+                    implementation(library(jetbrains_kotlinx_coroutines_core))
                 }
 
                 forKotlinSetsDependencies(
@@ -57,13 +57,13 @@ constructor(
                     TEST_INTEGRATION,
                     TEST_FIXTURES
                 ) {
-                    implementation(library(jetbrains_kotlinx_kotlinxCoroutinesTest))
+                    implementation(library(jetbrains_kotlinx_coroutines_test))
                 }
             }
         }
         configurable(isEnabled = isAnyAndroidFullEnabled) {
             forKotlinSetsDependencies(MAIN, ANDROID_MAIN) {
-                implementation(library(jetbrains_kotlinx_kotlinxCoroutinesAndroid))
+                implementation(library(jetbrains_kotlinx_coroutines_android))
             }
         }
     }